requirements.txt
certifi==2018.11.29
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.4.0
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
mysqlclient==1.3.14
psycopg2==2.7.6.1
python-dotenv==0.10.1
pytz==2018.7
whitenoise==4.1.2
wincertstore==0.2
